refactor Expected Errors
