+++ pages/docs/error-management/expected-errors.mdx
@@ -47,14 +47,14 @@ The following program serves as an illustration of how errors are automatically 
 <Tabs>
 <Tab>
 
-```ts filename="error-accumulation.ts" file=<rootDir>/src/guide/error-management/expected-errors/error-accumulation-gen.ts#L1-L27
+```ts filename="error-tracking.ts" file=<rootDir>/src/guide/error-management/expected-errors/error-tracking-gen.ts#L1-L27
 
 ```
 
 </Tab>
 <Tab>
 
-```ts filename="error-accumulation.ts" file=<rootDir>/src/guide/error-management/expected-errors/error-accumulation.ts#L1-L26
+```ts filename="error-tracking.ts" file=<rootDir>/src/guide/error-management/expected-errors/error-tracking.ts#L1-L23
 
 ```
 
@@ -105,25 +105,36 @@ Each operation depends on the successful execution of the previous one.
 If any error occurs, the execution is short-circuited, and the error is propagated.
 In this specific example, `operation3` is never executed because an error occurs in `operation2`.
 
-## Catching
-
-To handle errors in a simple way, you can use the `catchAll` function.
-
-### catchAll
-
-The `catchAll` function allows you to catch any error that occurs in the program and provide a fallback.
+## Catching all Errors
 
 <Tabs>
 <Tab>
 
-```ts {7-10} file=<rootDir>/src/guide/error-management/expected-errors/catchAll-gen.ts
+The [`Effect.either`](error-channel-operations#exposing-errors-in-the-success-channel) function converts an `Effect<R, E, A>` into another effect where both its failure (`E`) and success (`A`) channels have been lifted into an `Either<E, A>` data type:
+
+```ts
+Effect<R, E, A> -> Effect<R, never, Either<E, A>>
+```
+
+The [`Either<E, A>`](../data-types/either) data type represents a value that can be either an error (`E`) or a successful result (`A`).
+By yielding an `Either`, we gain the ability to **pattern match** on this type to handle both failure and success cases within the generator function.
+
+```ts /Effect.either/ file=<rootDir>/src/guide/error-management/expected-errors/catch-all-gen.ts
+
+```
+
+We can make the code less verbose by using the `Either.match` function, which directly accepts the two callback functions for handling errors and successful values:
+
+```ts file=<rootDir>/src/guide/error-management/expected-errors/catch-all-gen-match.ts
 
 ```
 
 </Tab>
 <Tab>
 
-```ts {6} file=<rootDir>/src/guide/error-management/expected-errors/catchAll-pipe.ts
+The `catchAll` function allows you to catch any error that occurs in the program and provide a fallback.
+
+```ts {6} file=<rootDir>/src/guide/error-management/expected-errors/catch-all-pipe.ts
 
 ```
 
@@ -137,30 +148,16 @@ indicating that all errors has been handled.
 Effect<never, never, string>
 ```
 
-### catchTag
+## Catching a Specific Error
 
-Suppose we only want to handle `FooError`. In that case, we can use the `catchTag` method provided by Effect:
+Suppose we want to handle a specific error, such as `FooError`.
 
 <Tabs>
 <Tab>
 
-```ts {9-11} file=<rootDir>/src/guide/error-management/expected-errors/catchTag-gen.ts
+```ts {9-11} file=<rootDir>/src/guide/error-management/expected-errors/catch-error-gen.ts
 
 ```
-
-</Tab>
-<Tab>
-
-```ts {6-8} file=<rootDir>/src/guide/error-management/expected-errors/catchTag-pipe.ts
-
-```
-
-In the example above, the `catchTag` function allows us to handle `FooError` specifically.
-If a `FooError` occurs during the execution of the program, the provided error handler function will be invoked,
-and the program will proceed with the recovery logic specified within the handler.
-
-</Tab>
-</Tabs>
 
 We can observe that the type in the error channel of our program has changed to only show `BarError`,
 indicating that `FooError` has been handled.
@@ -169,24 +166,43 @@ indicating that `FooError` has been handled.
 Effect<never, BarError, string>
 ```
 
-If we also wanted to handle `BarError`, we can simply add another `catchTag`.
+If we also want to handle `BarError`, we can easily add another case to our code:
 
-<Tabs>
-<Tab>
-
-```ts {12} file=<rootDir>/src/guide/error-management/expected-errors/catchTag-gen-2.ts
+```ts {12} file=<rootDir>/src/guide/error-management/expected-errors/catch-error-gen-2.ts
 
 ```
+
+We can observe that the type in the error channel of our program has changed to `never`,
+indicating that all errors has been handled.
 
 </Tab>
 <Tab>
 
-```ts {9-11} file=<rootDir>/src/guide/error-management/expected-errors/catchTag-pipe-2.ts
+If **all** the errors are **tagged** we can catch a specific error using the `Effect.catchTag` function:
+
+```ts {6-8} file=<rootDir>/src/guide/error-management/expected-errors/catch-error-pipe.ts
 
 ```
 
-</Tab>
-</Tabs>
+In the example above, the `Effect.catchTag` function allows us to handle `FooError` specifically.
+If a `FooError` occurs during the execution of the program, the provided error handler function will be invoked,
+and the program will proceed with the recovery logic specified within the handler.
+
+We can observe that the type in the error channel of our program has changed to only show `BarError`,
+indicating that `FooError` has been handled.
+
+```ts /BarError/
+Effect<never, BarError, string>
+```
+
+If we also wanted to handle `BarError`, we can simply add another `catchTag`:
+
+```ts {9-11} file=<rootDir>/src/guide/error-management/expected-errors/catch-error-pipe-2.ts
+
+```
+
+We can observe that the type in the error channel of our program has changed to `never`,
+indicating that all errors has been handled.
 
 <Warning>
   It is important to ensure that the error type used with `catchTag` has a
@@ -194,33 +210,21 @@ If we also wanted to handle `BarError`, we can simply add another `catchTag`.
   and handling of specific error tags.
 </Warning>
 
-### catchTags
+Instead of using the `Effect.catchTag` function multiple times to handle individual error types, we have a more convenient option called `Effect.catchTags`. With `Effect.catchTags`, we can handle multiple errors in a single block of code.
 
-Instead of using the `catchTag` combinator multiple times to handle individual error types, we have a more convenient option called `catchTags`. With `catchTags`, we can handle multiple errors in a single block of code.
-
-<Tabs>
-<Tab>
-
-```ts {9-14} file=<rootDir>/src/guide/error-management/expected-errors/catchTags-gen.ts
+```ts {6-9} file=<rootDir>/src/guide/error-management/expected-errors/catch-errors-pipe.ts
 
 ```
 
-</Tab>
-<Tab>
-
-```ts {6-9} file=<rootDir>/src/guide/error-management/expected-errors/catchTags-pipe.ts
-
-```
-
-In the above example, instead of using `catchTag` multiple times to handle individual errors, we utilize the `catchTags` combinator.
+In the above example, instead of using `Effect.catchTag` multiple times to handle individual errors, we utilize the `Effect.catchTags` combinator.
 This combinator takes an object where each property represents a specific error `_tag` (`"FooError"` and `"BarError"` in this case),
 and the corresponding value is the error handler function to be executed when that particular error occurs.
 
+<Warning>
+  It is important to ensure that all the error types used with
+  `Effect.catchTags` have a `readonly _tag` discriminant field. This field is
+  required for the matching and handling of specific error tags.
+</Warning>
+
 </Tab>
 </Tabs>
-
-<Warning>
-  It is important to ensure that all the error types used with `catchTags` have
-  a `readonly _tag` discriminant field. This field is required for the matching
-  and handling of specific error tags.
-</Warning>
