+++ src/guide/error-management/expected-errors/catch-all-pipe.ts
@@ -1,5 +1,5 @@
 import { Effect } from "effect"
-import { program } from "./error-accumulation"
+import { program } from "./error-tracking"
 
 // $ExpectType Effect<never, never, string>
 const recovered = program.pipe(
