+++ src/guide/error-management/expected-errors/catch-error-gen.ts
@@ -1,15 +1,16 @@
 import { Effect, Either } from "effect"
-import { program } from "./error-accumulation"
+import { program } from "./error-tracking"
 
 // $ExpectType Effect<never, BarError, string>
 const recovered = Effect.gen(function* (_) {
-  const successOrFailure = yield* _(Effect.either(program))
-  if (Either.isLeft(successOrFailure)) {
-    const error = successOrFailure.left
+  const failureOrSuccess = yield* _(Effect.either(program))
+  if (Either.isLeft(failureOrSuccess)) {
+    const error = failureOrSuccess.left
     if (error._tag === "FooError") {
       return "Recovering from FooError"
     }
     return yield* _(Effect.fail(error))
+  } else {
+    return failureOrSuccess.right
   }
-  return successOrFailure.right
 })
