--- src/guide/error-management/expected-errors/catchAll-gen.ts
@@ -1,12 +1,0 @@
-import { Effect, Either } from "effect"
-import { program } from "./error-accumulation"
-
-// $ExpectType Effect<never, never, string>
-const recovered = Effect.gen(function* (_) {
-  const successOrFailure = yield* _(Effect.either(program))
-  if (Either.isLeft(successOrFailure)) {
-    const error = successOrFailure.left
-    return `Recovering from ${error._tag}`
-  }
-  return successOrFailure.right
-})
